Create README.md best coffee Ethiopia
As a lover of coffee I wanted to find out the best quality of coffee in terms of flavour and Taste in Ethiopia so I looked for a dataset to help me with my research. I got a coffee quality dataset and the database was scrapped from Coffee Quality Institute in the year 2018. Firstly, I had to clean the dataset by finding the number of duplicates and removing the duplicates.  Secondly, I changed the column name processing_method to creation to make it shorter. Thirdly,  I filtered the dataset to  have only country Ethiopia,  fourthly, I added the column Flavor and Aftertaste to get the best_quality. Lastly, I looked for the number greater than 17 and the best quality coffee in Ethiopia is specie Arabica, owner metad plc, region Oromiya, best quality number 17.25